# shared/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8090
  description: Generated server url
paths:
...
  /api/leads:
    get:
      tags:
      - leads
      summary: Search/filter leads (marketing sees own; admins see all)
      operationId: list_1
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - LEAD
          - HOT_LEAD
          - APPLICATION
          - STUDENT
          - STALLED
      - name: countryCode
        in: query
        required: false
        schema:
          type: string
      - name: courseId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: channelId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: assignedTo
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: createdFrom
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: createdTo
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 25
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LeadPageDto"
    post:
      tags:
      - leads
      summary: Create a lead (owner defaults to the caller; admins may assign)
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LeadDto"
  /api/leads/{id}/transition:
    post:
      tags:
      - leads
      summary: Move a lead through the funnel; every accepted move writes an immutable
        event
      operationId: transition
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LeadDto"
...
  /api/leads/{id}:
    get:
      tags:
      - leads
      summary: Lead detail with full status history
      operationId: detail
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LeadDetailDto"
    patch:
      tags:
      - leads
      summary: Update lead data; reassignment is admin-only and audited
      operationId: update_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LeadDto"
...
  /api/leads/duplicates:
    get:
      tags:
      - leads
      summary: "Possible duplicates by email/phone — warn, don't block (RN-04)"
      operationId: duplicates
      parameters:
      - name: email
        in: query
        required: false
        schema:
          type: string
      - name: phone
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DuplicateDto"
...
  /api/analytics/summary:
    get:
      tags:
      - analytics
      summary: Headline KPIs
      operationId: summary
      parameters:
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Summary"
...
components:
  schemas:
...
    CreateLeadRequest:
      type: object
      properties:
        fullName:
          type: string
          minLength: 1
        countryCode:
          type: string
          minLength: 1
          pattern: "[A-Za-z]{2}"
        email:
          type: string
        phone:
          type: string
        courseId:
          type: string
          format: uuid
        channelId:
          type: string
          format: uuid
        assignedTo:
          type: string
          format: uuid
        utmSource:
          type: string
        utmMedium:
          type: string
        utmCampaign:
          type: string
      required:
      - channelId
      - countryCode
      - courseId
      - fullName
    LeadDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        countryCode:
          type: string
        email:
          type: string
        phone:
          type: string
        courseId:
          type: string
          format: uuid
        channelId:
          type: string
          format: uuid
        utmSource:
          type: string
        utmMedium:
          type: string
        utmCampaign:
          type: string
        status:
          type: string
          enum:
          - LEAD
          - HOT_LEAD
          - APPLICATION
          - STUDENT
          - STALLED
        stalledFromStatus:
          type: string
          enum:
          - LEAD
          - HOT_LEAD
          - APPLICATION
          - STUDENT
          - STALLED
        stallReasonId:
          type: string
          format: uuid
        assignedTo:
          type: string
          format: uuid
        assignedToName:
          type: string
        lastContactedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    TransitionRequest:
      type: object
      properties:
        toStatus:
          type: string
          enum:
          - LEAD
          - HOT_LEAD
          - APPLICATION
          - STUDENT
          - STALLED
        stallReasonId:
          type: string
          format: uuid
        note:
          type: string
      required:
      - toStatus
...
    UpdateLeadRequest:
      type: object
      properties:
        fullName:
          type: string
        countryCode:
          type: string
          pattern: "[A-Za-z]{2}"
        email:
          type: string
        phone:
          type: string
        courseId:
          type: string
          format: uuid
        channelId:
          type: string
          format: uuid
        assignedTo:
          type: string
          format: uuid
...
    LeadPageDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/LeadDto"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
    LeadDetailDto:
      type: object
      properties:
        lead:
          $ref: "#/components/schemas/LeadDto"
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/StatusEventDto"
    StatusEventDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fromStatus:
          type: string
          enum:
          - LEAD
          - HOT_LEAD
          - APPLICATION
          - STUDENT
          - STALLED
        toStatus:
          type: string
          enum:
          - LEAD
          - HOT_LEAD
          - APPLICATION
          - STUDENT
          - STALLED
        stallReasonId:
          type: string
          format: uuid
        note:
          type: string
        changedBy:
          type: string
          format: uuid
        changedByName:
          type: string
        changedAt:
          type: string
          format: date-time
    DuplicateDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        status:
          type: string
          enum:
          - LEAD
          - HOT_LEAD
          - APPLICATION
          - STUDENT
          - STALLED
        assignedToName:
          type: string
...
    Summary:
      type: object
      properties:
        totalLeads:
          type: integer
          format: int64
        newLeads:
          type: integer
          format: int64
        conversionRate:
          type: number
          format: double
        avgDaysToConvert:
          type: number
          format: double
        slaBreaches:
          type: integer
          format: int64
...